♻️ Refactor : atoms ui
