+++ src/feature/picsel/shared/components/ui/molecules/FloatingActionButtons.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 import { View } from 'react-native';
 
-import FunctionButton from '../atoms/FunctionButton';
+import FunctionButton from '../atoms/Button/FunctionButton';
 
-import FloatingAddButton from '@/feature/picsel/shared/components/ui/atoms/AddButton';
-import UpButton from '@/feature/picsel/shared/components/ui/atoms/UpButton';
+import FloatingAddButton from '@/feature/picsel/shared/components/ui/atoms/Button/AddButton';
+import UpButton from '@/feature/picsel/shared/components/ui/atoms/Button/UpButton';
 
 interface FloatingActionButtonsProps {
   isSelecting: boolean;
