+++ src/feature/picsel/shared/components/ui/molecules/SelectablePhotoCard.tsx
@@ -4,7 +4,7 @@ import { Image, Pressable, Text, View } from 'react-native';
 
 import { formatDate } from '../../../utils/dateUtils';
 
-import SelectionCheckbox from '@/feature/picsel/shared/components/ui/atoms/SelectionCheckbox';
+import SelectionCheckbox from '@/feature/picsel/myPicsel/components/ui/atoms/SelectionCheckbox';
 import SparkleImages from '@/shared/images/Sparkle';
 
 interface Props {
