+++ src/feature/picsel/shared/components/ui/template/myPicsel/MyPicselTemplate.tsx
@@ -3,10 +3,12 @@ import React, { useEffect, useState } from 'react';
 import { useNavigation } from '@react-navigation/native';
 import { View } from 'react-native';
 
-import FloatingAddButton from '../../atoms/AddButton';
-import AddButton from '../../atoms/AddButton';
-import DateFilterButton, { DateFilterType } from '../../atoms/DateFilterButton';
-import FunctionButton from '../../atoms/FunctionButton';
+import DateFilterButton, {
+  DateFilterType,
+} from '../../../../../myPicsel/components/ui/atoms/DateFilterButton';
+import FloatingAddButton from '../../atoms/Button/AddButton';
+import AddButton from '../../atoms/Button/AddButton';
+import FunctionButton from '../../atoms/Button/FunctionButton';
 import EmptyStateLayout from '../../layouts/EmptyStateLayout';
 import EmptyMessage from '../../molecules/EmptyMessage';
 import UploadTooltip from '../../molecules/UploadTooltip';
@@ -17,7 +19,7 @@ import MonthFilterView from '@/feature/picsel/myPicsel/components/ui/organisms/M
 import PhotoListView from '@/feature/picsel/myPicsel/components/ui/organisms/PhotoListView';
 import YearFilterView from '@/feature/picsel/myPicsel/components/ui/organisms/YearFilterView';
 import { MOCK_YEAR_DATA } from '@/feature/picsel/myPicsel/data/MOCK_YEAR_DATA';
-import UpButton from '@/feature/picsel/shared/components/ui/atoms/UpButton';
+import UpButton from '@/feature/picsel/shared/components/ui/atoms/Button/UpButton';
 import { useFunctionButtons } from '@/feature/picsel/shared/hooks/useFunctionButtons';
 import { usePhotoActions } from '@/feature/picsel/shared/hooks/usePhotoActions';
 import { usePhotoSelection } from '@/feature/picsel/shared/hooks/usePhotoSelection';
